% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_serve.R
\name{calibrate_serve}
\alias{calibrate_serve}
\title{Calibrates Serve Probabilities}
\usage{
calibrate_serve(win, player.serve, opponent.serve, bestof3 = TRUE)
}
\arguments{
\item{win}{Numeric value between 0 and 1 that is the player's expected win percentage}

\item{bestof3}{Logical indicating whether the match is a best of 3 (TRUE) or best of 5 (FALSE) format}

\item{player.serve.}{Numerical expected serve}

\item{opponent.serve.}{Numerical expected serve}
}
\description{
Returns the expected probability of player given that player's win expectations
}
