% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_win_dynamic.R
\name{dynamic_in_match_win}
\alias{dynamic_in_match_win}
\title{Dynamic Match Win Prediction}
\usage{
dynamic_in_match_win(point_a, point_b, game_a, game_b, set_a, set_b,
  server.prob, returner.prob, server.serve.points.won, server.serve.points,
  returner.serve.points.won, returner.serve.points, bestof3 = TRUE)
}
\arguments{
\item{point_a}{Numeric game points won by current server at start of point}

\item{point_b}{Numeric game points won by current returner at start of point}

\item{game_a}{Numeric games won by current server in the current set}

\item{game_b}{Numeric games won by current returner in the current set}

\item{set_a}{Numeric sets won by current server}

\item{set_b}{Numeric sets won by current returner}

\item{server.prob}{Numeric serve win prob of current server}

\item{server.serve.points.won}{Numeric total service points won in match up to current point}

\item{server.serve.points}{Numeric total service points played in match up to current point}

\item{returner.serve.points.won}{Numeric total service points won in match up to current point}

\item{returner.serve.points}{Numeric total service points played in match up to current point}

\item{return.prob}{Numeric serve win prob of current returner}
}
\description{
This function calculates the match win prediction for matches in progress based on the current score and the in-match server performance
}
