% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_lookup.R
\name{elo_lookup}
\alias{elo_lookup}
\title{Current Elo}
\usage{
elo_lookup(id, mens)
}
\arguments{
\item{id}{Character id of player}

\item{mens}{Logical if mens or womens match}
}
\value{
Numeric elo value
}
\description{
Extracts current elo from stored "elo" data
}
