% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{get_baseline}
\alias{get_baseline}
\title{Get Baseline Serve and Win Probabilities}
\usage{
get_baseline(playerid, opponentid, mens = TRUE, format)
}
\arguments{
\item{playerid.}{4-character player id}

\item{opponentid.}{4-character opponent id}

\item{mens.}{Logical TRUE if mens match, FALSE if womens}

\item{format.}{Character ('bestof3', 'laver' or 'bestof5')}
}
\value{
data frame of player 1 and player 2 win probs
}
\description{
Gets players current chances of winning point based on score at end
}
