% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_win_prob.R, R/winprob.R
\name{get_win_prob}
\alias{get_win_prob}
\alias{get_win_prob}
\title{Win Prob at End of Point}
\usage{
get_win_prob(playerid, opponentid, mens = TRUE)

get_win_prob(playerid, opponentid, mens = TRUE)
}
\arguments{
\item{mens}{Logical if men's match}

\item{data}{Data frame from \code{point_data}}

\item{data}{Data frame from \code{point_data}}

\item{mens}{Logical if men's match}
}
\value{
data frame of player 1 and player 2 win probs

data frame of player 1 and player 2 win probs
}
\description{
Gets players current chances of winning point based on score at end

Gets players current chances of winning point based on score at end
}
