% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_lookup.R
\name{serve_lookup}
\alias{serve_lookup}
\title{Lookup Barnett and Clark Serve}
\usage{
serve_lookup(player1, player2)
}
\arguments{
\item{Character}{of player1 name}

\item{Character}{of player2 name}
}
\description{
Lookup Barnett and Clark Serve
}
