% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_prior_lookup.R
\name{serve_prior_lookup}
\alias{serve_prior_lookup}
\title{Get Serve Priors}
\usage{
serve_prior_lookup(
  player_id,
  player2_id = NULL,
  opponent_id,
  opponent2_id = NULL,
  mens
)
}
\arguments{
\item{player_id.}{ATP/WTA id for player}

\item{player2_id.}{ATP/WTA id for player (only for doubles)}

\item{opponent_id.}{ATP/WTA id for player}

\item{opponent2_id.}{ATP/WTA id for player (only for doubles)}

\item{mens.}{Logical if ATP or WTA match}
}
\value{
list of expected serve probabilities
}
\description{
Opponent corrected serve expectation
}
