% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_score.R
\name{update_score}
\alias{update_score}
\title{Update Score with Advantage Final Set}
\usage{
update_score(pointa, pointb, gamea, gameb, seta, setb, bestof3 = T)
}
\arguments{
\item{bestof3}{Logical indicator if best-of-3 match}

\item{point_a}{Numeric game points won by current server at start of point}

\item{point_b}{Numeric game points won by current returner at start of point}

\item{game_a}{Numeric games won by current server in the current set}

\item{game_b}{Numeric games won by current returner in the current set}

\item{set_a}{Numeric sets won by current server}

\item{set_b}{Numeric sets won by current returner}
}
\description{
Check score change based on result of current point and return score in terms of serving player of the previous point
}
