% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_score_laver.R
\name{update_score_laver}
\alias{update_score_laver}
\title{Update Score Laver}
\usage{
update_score_laver(pointa, pointb, gamea, gameb, seta, setb)
}
\arguments{
\item{point_a}{Numeric game points won by current server at start of point}

\item{point_b}{Numeric game points won by current returner at start of point}

\item{game_a}{Numeric games won by current server in the current set}

\item{game_b}{Numeric games won by current returner in the current set}

\item{set_a}{Numeric sets won by current server}

\item{set_b}{Numeric sets won by current returner}
}
\description{
Check score change based on result of current point for Laver Cup format
}
