% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_lookup.R
\name{elo_lookup}
\alias{elo_lookup}
\title{Current Elo}
\usage{
elo_lookup(id1, id2 = NULL, mens, default = 1300)
}
\arguments{
\item{id1}{Character id of player 1}

\item{id2}{Character id of player 2 (if doubles)}

\item{mens}{Logical if mens or womens match}

\item{default}{Numerical rating used when ID not found}
}
\value{
Numeric elo value
}
\description{
Extracts current elo from stored "elo" data
}
